+++ packages/migrate/tests/fixtures/migrate/output/complex.ts.output
@@ -1,16 +1,16 @@
 import { salutations, DEFAULT_GREETING, Greeting } from './salutations';
 
-export function hasGreeting(g: Greeting): Greeting {
+export function hasGreeting(g: Greeting) {
   return salutations.find((s) => s == g) || DEFAULT_GREETING;
 }
 
-export function findGreeting(someGreeting: Greeting): Greeting {
+export function findGreeting(someGreeting: Greeting) {
   const greeting = salutations.find((g) => {
     return g == someGreeting;
   });
   return greeting;
 }
 
-export default function say(name = 'World'): string {
+export default function say(name = 'World') {
   return `Hello ${name}`;
 }
