+++ packages/migrate/tests/fixtures/migrate/output/index.ts.output
@@ -1,3 +1,3 @@
-export function split(str: string): string[] {
+export function split(str: string) {
   return str.split(',');
 }
