+++ packages/plugins/src/index.ts
@@ -1,3 +1,2 @@
-export * from './fix-transform';
 export * from './plugins';
 export * from './data';
