+++ packages/plugins/src/plugins/index.ts
@@ -3,4 +3,4 @@ export { EmptyLinesRestorePlugin } from './empty-lines-restore.plugin';
 export { LintPlugin } from './lint.plugin';
 export { ReRehearsePlugin } from './re-rehearse.plugin';
 export { DiscoverTypesPlugin } from './discover-types.plugin';
-export { DiagnosticAutofixPlugin } from './diagnostics-autofix.plugin';
+export { DiagnosticAutofixPlugin } from './diagnostic-autofix.plugin';
