+++ packages/transforms/src/7006/fixtures/7006.ts.output
@@ -1,3 +1,3 @@
-export function split(str: string): string[] {
+export function split(str: string) {
   return str.split(',');
 }
