+++ packages/transforms/src/index.ts
@@ -1,2 +1,3 @@
 export * from './codefix-collection';
 export * from './codefixes';
+export * from './fix-transform';
