+++ packages/upgrade/test/transforms.test.ts
@@ -1,12 +1,11 @@
 import fs from 'fs';
 import { afterEach, describe, expect, test } from 'vitest';
 import { parse, resolve } from 'path';
-
 import { upgrade } from '../src';
 
 describe('Test transform', function () {
   const basePath = resolve(__dirname, 'fixtures', 'transforms');
-  const transformsPath = resolve(__dirname, '..', 'src', 'transforms');
+  const transformsPath = resolve(__dirname, '../../transforms/src');
   const transforms = fs
     .readdirSync(transformsPath, { withFileTypes: true })
     .filter((file) => file.isDirectory())
